+++ src/fcn/updater/2_x/2_10.php
@@ -1,0 +1,35 @@
+<?php
+/**
+ * Acuparse - AcuRite®‎ Access/smartHUB and IP Camera Data Processing, Display, and Upload.
+ * @copyright Copyright (C) 2015-2019 Maxwell Power
+ * @author Maxwell Power <max@acuparse.com>
+ * @link http://www.acuparse.com
+ * @license AGPL-3.0+
+ *
+ * This code is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU Affero General Public License as published
+ * by the Free Software Foundation, either version 3 of the License, or
+ * any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
+ * GNU Affero General Public License for more details.
+ *
+ * You should have received a copy of the GNU Affero General Public License
+ * along with this code. If not, see <http://www.gnu.org/licenses/>.
+ */
+
+/**
+ * File: src/fcn/updater/2_x/2_10.php
+ * 2.10 Site Update Tasks
+ */
+
+switch ($config->version->app) {
+
+    // Update from 2.9.4
+    case '2.9.4-release':
+        $config->upload->myacurite->pass_unknown = false;
+        $config->version->app = '2.10.0-release';
+        $notes .= '<li><strong>' . $config->version->app . '</strong> - ' . 'Support Lightning Tower & Upload Unknown Sensors';
+}