+++ src/usr/config.new.php
@@ -160,7 +160,8 @@ return (object)array(
         // MyAcurite
         'myacurite' => (object)array(
             'access_enabled' => true, // true or false
-            'access_url' => 'https://atlasapi.myacurite.com',
+            'access_url' => 'https://atlasapi.myacurite.com', //MyAcurite API
+            'pass_unknown' => false, // Pass unadded sensors to MyAcurite?
         ),
     ),
 
@@ -194,7 +195,7 @@ return (object)array(
 
     // Application/Database Version
     'version' => (object)array(
-        'app' => '2.9.4-release',
+        'app' => '2.10.0-release',
         'schema' => '2.9',
     ),
 );